use generated exprs to test expression evaluation against smt solver
