+++ src/test/scala/BitVectorSMTTest.scala
@@ -1,12 +1,18 @@
 import ir.*
+import org.scalacheck.{Arbitrary, Gen}
 import org.scalatest.*
 import org.scalatest.funsuite.*
+import test_util.{CaptureOutput, ExprGen}
 import translating.BasilIRToSMT2
 import translating.PrettyPrinter.*
+import util.Logger
 import util.z3.*
 
 @test_util.tags.UnitTest
-class BitVectorEvalTest extends AnyFunSuite {
+class BitVectorEvalTest
+    extends AnyFunSuite
+    with org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
+    with CaptureOutput {
 
   def genSMT(size: Int) = {
     val max = BitVecType(size).maxValue.toInt
@@ -67,5 +73,20 @@ class BitVectorEvalTest extends AnyFunSuite {
     }
   }
 
+  implicit lazy val arbExpr: Arbitrary[Expr] = Arbitrary(for {
+    sz <- Gen.oneOf(List(30, 31, 32, 33, 60, 63, 64, 65, 66, 70, 90, 128, 126))
+    e <- ExprGen.genExpr(Some(sz))
+  } yield (e))
+
+  test("interp exprs smt") {
+    forAll(minSuccessful(30)) { (exp: Expr) =>
+      val test = BinaryExpr(NEQ, exp, ir.eval.evaluateExpr(exp).get)
+      val q = BasilIRToSMT2.exprUnsat(test, None, false)
+      Logger.info("assert: " + test)
+      util.z3.checkSATSMT2(q, Some(30000)) == SatResult.UNSAT
+    }
+  }
+
   genSMT(2)
+
 }
